frigate/requirements.txt
meshtastic
requests
paho-mqtt
